chore(folio): extract zoom and page-indicator state into a hook
The `DocxEditor` component owned two intertwined concerns — zoom level
with viewport-anchor capture/replay, and the Google-Docs-style page
indicator driven by scroll position. Both lived inline alongside ~16
other state slices and contributed several effects, refs, and
callbacks. Move them into a new `useZoomAndPageInfo` hook so the editor
body shrinks (-126 lines net) and the zoom/page-info lifecycle is
testable in isolation.

The hook owns: `zoom` state (previously a field of the kitchen-sink
`EditorState`), the `zoomRef` mirror, the pending viewport anchor,
`scrollPageInfo` state, the fade timer, and the three effects that
keep them in sync (mirror zoom into `zoomRef`, apply the captured
anchor after a reflow, attach a scroll listener once the container
mounts). `state.zoom` is removed from `EditorState` — the editor now
reads `zoom` directly from the hook.

While here, the throwaway `handleZoomChange` wrapper around
`setZoomWithViewportAnchor` is removed; the toolbar binds directly to
the setter. The two magic numbers in the scroll-page-info math
(`pageGap = 24`, `paddingTop = 24`, fade `600` ms) become named
constants inside the hook.

No behavior change.
